refactor server, middleware, and Schedule fetch
